Fixing draw call float->int cast issue

+++ src/main/java/com/openbravo/pos/printer/ticket/PrintItemLine.java
@@ -83,11 +83,11 @@ public class PrintItemLine implements PrintItem {
     public void draw(Graphics2D g, int x, int y, int width) {
 
         MyPrinterState ps = new MyPrinterState(textsize);
-        float left = x;
+        int left = x;
         for (int i = 0; i < m_atext.size(); i++) {
             StyledText t = m_atext.get(i);
             g.setFont(ps.getFont(font, t.style));
-            g.drawString(t.text, left, (float) y);
+            g.drawString(t.text, left, y);
             left += g.getFontMetrics().getStringBounds(t.text, g).getWidth();
         }
     }
